# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "University Ecosystem API",
    "description": "REST API for university life management platform - schedules, news, events, notifications, and more.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/mfa/totp/start": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Start Totp Enrollment Endpoint",
        "operationId": "start_totp_enrollment_endpoint_api_v1_auth_mfa_totp_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TotpEnrollmentStartIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpEnrollmentStartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/totp/confirm": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Confirm Totp Enrollment",
        "operationId": "confirm_totp_enrollment_api_v1_auth_mfa_totp_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpEnrollmentConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaTotpEnrollmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/totp": {
      "get": {
        "tags": [
          "mfa"
        ],
        "summary": "List Totp Enrollments",
        "operationId": "list_totp_enrollments_api_v1_auth_mfa_totp_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MfaTotpEnrollmentOut"
                  },
                  "type": "array",
                  "title": "Response List Totp Enrollments Api V1 Auth Mfa Totp Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/totp/pending/{enrollment_id}": {
      "delete": {
        "tags": [
          "mfa"
        ],
        "summary": "Delete Pending Totp Enrollment",
        "operationId": "delete_pending_totp_enrollment_api_v1_auth_mfa_totp_pending__enrollment_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "enrollment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Enrollment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/totp/{enrollment_id}": {
      "delete": {
        "tags": [
          "mfa"
        ],
        "summary": "Delete Totp Enrollment",
        "operationId": "delete_totp_enrollment_api_v1_auth_mfa_totp__enrollment_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "enrollment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Enrollment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaFactorStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/webauthn/register/start": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Start Webauthn Registration",
        "operationId": "start_webauthn_registration_api_v1_auth_mfa_webauthn_register_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnRegistrationOptionsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/webauthn/register/confirm": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Confirm Webauthn Registration",
        "operationId": "confirm_webauthn_registration_api_v1_auth_mfa_webauthn_register_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnRegistrationVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaFactorStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/webauthn": {
      "get": {
        "tags": [
          "mfa"
        ],
        "summary": "List Webauthn Credentials",
        "operationId": "list_webauthn_credentials_api_v1_auth_mfa_webauthn_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Webauthn Credentials Api V1 Auth Mfa Webauthn Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/webauthn/{credential_id}": {
      "delete": {
        "tags": [
          "mfa"
        ],
        "summary": "Delete Webauthn Credential",
        "operationId": "delete_webauthn_credential_api_v1_auth_mfa_webauthn__credential_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credential Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaFactorStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/recovery-codes": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Generate Recovery Codes Endpoint",
        "operationId": "generate_recovery_codes_endpoint_api_v1_auth_mfa_recovery_codes_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesGenerateOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/step-up": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Request Step Up",
        "operationId": "request_step_up_api_v1_auth_mfa_step_up_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingMfaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MfaFactorStatusOut": {
        "properties": {
          "disabled": {
            "type": "boolean",
            "title": "Disabled"
          },
          "mfa_default_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mfa Default Method"
          },
          "mfa_required": {
            "type": "boolean",
            "title": "Mfa Required",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "disabled"
        ],
        "title": "MfaFactorStatusOut"
      },
      "MfaMethodChallengeOut": {
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "totp",
              "webauthn"
            ],
            "title": "Method"
          },
          "challenge_token": {
            "type": "string",
            "title": "Challenge Token"
          },
          "challenge_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Challenge Expires At"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "attempt_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Count"
          },
          "attempt_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Limit"
          },
          "remaining_attempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Attempts"
          }
        },
        "type": "object",
        "required": [
          "method",
          "challenge_token",
          "challenge_expires_at"
        ],
        "title": "MfaMethodChallengeOut"
      },
      "MfaTotpEnrollmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "is_active",
          "created_at"
        ],
        "title": "MfaTotpEnrollmentOut"
      },
...
      "PendingMfaResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "mfa_required",
            "title": "Status",
            "default": "mfa_required"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "default_method": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "totp",
                  "webauthn"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Method"
          },
          "methods": {
            "items": {
              "$ref": "#/components/schemas/MfaMethodChallengeOut"
            },
            "type": "array",
            "title": "Methods"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "methods"
        ],
        "title": "PendingMfaResponse"
      },
...
      "RecoveryCodesGenerateOut": {
        "properties": {
          "codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Codes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "codes",
          "created_at"
        ],
        "title": "RecoveryCodesGenerateOut"
      },
...
      "TotpEnrollmentConfirmIn": {
        "properties": {
          "enrollment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Enrollment Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "enrollment_id",
          "code"
        ],
        "title": "TotpEnrollmentConfirmIn"
      },
      "TotpEnrollmentStartIn": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "reuse_existing": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reuse Existing",
            "default": false
          }
        },
        "type": "object",
        "title": "TotpEnrollmentStartIn"
      },
      "TotpEnrollmentStartOut": {
        "properties": {
          "enrollment": {
            "$ref": "#/components/schemas/MfaTotpEnrollmentOut"
          },
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "otpauth_url": {
            "type": "string",
            "title": "Otpauth Url"
          }
        },
        "type": "object",
        "required": [
          "enrollment",
          "secret",
          "otpauth_url"
        ],
        "title": "TotpEnrollmentStartOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebAuthnRegistrationOptionsOut": {
        "properties": {
          "publicKey": {
            "additionalProperties": true,
            "type": "object",
            "title": "Publickey"
          },
          "challenge_token": {
            "type": "string",
            "title": "Challenge Token"
          }
        },
        "type": "object",
        "required": [
          "publicKey",
          "challenge_token"
        ],
        "title": "WebAuthnRegistrationOptionsOut"
      },
      "WebAuthnRegistrationVerifyIn": {
        "properties": {
          "challenge": {
            "type": "string",
            "title": "Challenge"
          },
          "response": {
            "additionalProperties": true,
            "type": "object",
            "title": "Response"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "challenge",
          "response"
        ],
        "title": "WebAuthnRegistrationVerifyIn"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}